pyproject.toml
[project]
name = "vqm-ps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.5",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "boto3>=1.42.88",
    "botocore>=1.42.88",
    "cryptography>=46.0.7",
    "email-validator>=2.3.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "langchain>=0.3",
    "langchain-aws>=1.4.3",
    "langchain-community>=0.4.1",
    "langgraph>=0.2",
    "langsmith>=0.7.30",
    "msal>=1.36.0",
    "openai>=2.31.0",
    "openpyxl>=3.1.5",
    "opentelemetry-api>=1.41.0",
    "opentelemetry-exporter-otlp>=1.41.0",
    "opentelemetry-instrumentation-fastapi>=0.62b0",
    "opentelemetry-sdk>=1.41.0",
    "orjson>=3.11.8",
    "pdfplumber>=0.11.9",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.0",
    "pydantic-settings>=2.13.1",
    "pysnow>=0.7.17",
    "python-dateutil>=2.9.0.post0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "simple-salesforce>=1.12.9",
    "sqlalchemy[asyncio]>=2.0.49",
    "sshtunnel>=0.4.0",
    "structlog>=25.5.0",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.44.0",
]

[tool.pytest.ini_options]
pythonpath = [".", "src"]

[tool.ruff]
src = [".", "src"]

[dependency-groups]
dev = [
    "moto>=5.1.22",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.10",
]
